created and integrated Personalities component in the chatOnboarding component
